Refactor code to prep for tests
