+++ wallchart/static/fuse.js
