+++ wallchart/templates/admin.html
