+++ wallchart/templates/includes/message.html
