+++ wallchart/templates/worker.html
